# docs/swagger.yaml
basePath: /
definitions:
  controllers.BadRequestResp:
    properties:
      error:
        example: invalid request
        type: string
    type: object
  controllers.CreateAPIKeyResp:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
      secret:
        $ref: '#/definitions/models.AccountSecret'
    type: object
  controllers.DeleteAPIKeyReq:
    properties:
      secret_id:
        type: integer
    type: object
...
  controllers.GetAPIKeyResp:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
      secrets:
        items:
          $ref: '#/definitions/models.AccountSecret'
        type: array
    type: object
  controllers.RegistAccountReq:
    properties:
      avatar:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    type: object
...
  controllers.ReqRefreshToken:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  controllers.ReqUpdateProfile:
    properties:
      avatar_url:
        type: string
      description:
        type: string
      username:
        type: string
    type: object
  controllers.ServerErrorResp:
    properties:
      error:
        example: db error
        type: string
    type: object
  controllers.SuccessResp:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
    type: object
...
  models.AccountSecret:
    properties:
      accountID:
        type: integer
      apiKey:
        type: string
      createdAt:
        type: string
      expiredAt:
        type: string
      id:
        type: integer
      isDeleted:
        type: integer
      updatedAt:
        type: string
    type: object
info:
  contact:
    url: https://github.com/0xbarchitect
  description: All-in-one SSO Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSO
  version: "1.0"
paths:
...
  /create-api-key:
    post:
      operationId: CreateAPIKey
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.CreateAPIKeyResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.BadRequestResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ServerErrorResp'
      security:
      - JwtHeader: []
      summary: Create API Key
  /delete-api-key:
    delete:
      operationId: DeleteAPIKey
      parameters:
      - description: json body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/controllers.DeleteAPIKeyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.BadRequestResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ServerErrorResp'
      security:
      - JwtHeader: []
      summary: Delete API Key
...
  /get-api-key:
    get:
      operationId: GetAPIKey
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetAPIKeyResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.BadRequestResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ServerErrorResp'
      security:
      - JwtHeader: []
      summary: Get API Key
  /get_dareid_by_wallet_address/{wallet_address}:
    get:
      deprecated: true
      operationId: GetDareIdByWalletAddress
      parameters:
      - description: wallet address
        in: path
        name: wallet_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ClientBasicAuth: []
      summary: GetDareIdByWalletAddress
  /get_profile:
    get:
      operationId: GetProfile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JwtHeader: []
      summary: GetProfile
  /get_profile_by_dareid/:dareid:
    get:
      operationId: GetProfileByDareid
      parameters:
      - description: dareid
        in: path
        name: dareid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ClientBasicAuth: []
      summary: GetProfileByDareid
...
  /refresh_token:
    post:
      deprecated: true
      operationId: RefreshToken
      parameters:
      - description: refresh token req body
        in: body
        name: _
        schema:
          $ref: '#/definitions/controllers.ReqRefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ClientBasicAuth: []
      summary: RefreshToken - this API is deprecated, please use /oauth/token instead
  /regist-account:
    post:
      operationId: RegistAccountHandler
      parameters:
      - description: json body
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/controllers.RegistAccountReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: RegistAccountHandler
...
  /update_profile:
    post:
      operationId: UpdateProfile
      parameters:
      - description: update profile json body
        in: body
        name: _
        schema:
          $ref: '#/definitions/controllers.ReqUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JwtHeader: []
      summary: UpdateProfile
...
securityDefinitions:
  ClientBasicAuth:
    type: basic
  JwtHeader:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"